+++ src/component/RestaurantListing/Category/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CheckBox from '../../../common/checkbox';
+import CheckBox from '../../common/Checkbox';
 
 import './styles.scss'
 
@@ -8,7 +8,7 @@ const Category = (props) => {
   const list = [1, 2] || props.category;
   const checkBoxList = list.map(category => <CheckBox />);
 
-  return ( 
+  return (
     <div className="category">
       <div>Category</div>
       {checkBoxList}
