+++ src/component/RestaurantListing/Filters/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Slider from './Slider/index';
-import Category from './Category/index';
+import Category from '../Category/index';
 
 import './styles.scss';
 
@@ -8,8 +8,6 @@ const Filters = (props) => {
 
   return (
     <div className="filters">
-      <Category /> 
-      <br/>
       <Slider />
     </div>
    );
