+++ src/component/RestaurantListing/RestaurantSearchResult/styles.scss
