+++ src/component/RestaurantListing/styles.scss
@@ -1,5 +1,5 @@
-.listings {
-  display: flex;
-  flex-direction: row;
-  height: inherit;
-}
+// .listings {
+//   display: flex;
+//   flex-direction: row;
+//   height: inherit;
+// }
